Rendering in a separated file
