+++ main.py
@@ -3,33 +3,34 @@ import camera
 import engine
 import mapgen
 import generate_map
-import mouse
 import pygame
+import render
+import input
+
 WIN = pygame.display.set_mode((1920, 1080))
 CAM = camera.Camera(25, 10)
 SCALE = 60
 
-pygame.display.set_caption("rendkívül keratív név")
-icon = pygame.image.load('Assets/icon/icon.png')
-pygame.display.set_icon(icon)
-
 
 def main() -> None:
+    pygame.display.set_caption("rendkívül keratív név")
+    icon = pygame.image.load('Assets/icon/icon.png')
+    pygame.display.set_icon(icon)
     generate_map.generate_map(SCALE).generation()
     map: mapgen.Map = mapgen.Map(300, 1)
     map.generateTiles(SCALE)
     map.tile_set_colorkey()
     CAM.set_offset_to_middle()
-    run = True
     pos: tuple(int) = (0, 0)
     zoom_hud: hud.zoom = hud.zoom(CAM.zoom)
     zoom_in = False
     zoom_out = False
     clock = pygame.time.Clock()
+    run = True
     while run:
         DISPLAY = pygame.Surface((1920 + CAM.zoom[0], 1080 + CAM.zoom[1]))
         DISPLAY.fill((76.4, 107.3, 121.7))
-        mapgen.Map.renderTiles(CAM.move_camera(), DISPLAY)
+        render.renderTiles(CAM.move_camera(), DISPLAY)
         map.tiles = []
         clock.tick(60)
         pos = engine.MoveTowards(pos, (1820, 980), 20)
